Ajuste para ejercicios EXTRA

+++ src/P1.2 - Ejercicios básicos/ej09.py
@@ -1,5 +1,5 @@
 print("La suma de los tres números es:", 
-      int(input("Introduce el primer número: ")) + 
-      int(input("Introduce el segundo número: ")) + 
-      int(input("Introduce el tercer número: ")))
+      float(input("Introduce el primer número: ")) + 
+      float(input("Introduce el segundo número: ")) + 
+      float(input("Introduce el tercer número: ")))
